Move more transform based functions to the separate module
